refactor tron.js
- clean up variables and attributes
- create showWinner function
- create playerSpeed function

I added and additional push to the array of wall
coords. When the cycles are traveling at the
default speed of 1 the collision check works
correctly.

When the turbo feature was added the speed was
doubled. This meant that a solid wall was not
being stored the any cycles that were using
turbo, and it also gave a 50/50 chance that
the cycle would pass thru a wall and 'miss' the
collision because it was moving to every other
coord. 

I added an extra coord to the wall array. We now
push the current coord, and the previous coord.
This probably needs to be looked at to remove
duplicate coords, or changed to an object search.